One file for analysis
A bigger change to the analysis pipeline. A centralized analysis routine that calls different functions required.

Recordings should be saved in:
animal ID/Cell#

Execute gridAnalysis.m for the basic grid stimulation experiment analysis. Later the results of this analysis and of imaging results will be utilized by asymmetryCalc.m to compute a measure of asymmetry in the dendritic arbour of the cell.